requirements.txt
streamlit
yfinance
pandas
numpy
scikit-learn
umap-learn
plotly